Unshim join execs
