--- sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuBroadcastNestedLoopJoinExec.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2020-2021, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nvidia.spark.rapids
-
-import com.nvidia.spark.rapids.shims.v2._
-
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.plans.JoinType
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.execution.joins.BroadcastNestedLoopJoinExec
-import org.apache.spark.sql.rapids.execution._
-
-case class GpuBroadcastNestedLoopJoinExec(
-    left: SparkPlan,
-    right: SparkPlan,
-    join: BroadcastNestedLoopJoinExec,
-    joinType: JoinType,
-    condition: Option[Expression],
-    targetSizeBytes: Long)
-  extends GpuBroadcastNestedLoopJoinExecBase(left, right, joinType, condition,
-    targetSizeBytes) {
-
-  def getGpuBuildSide: GpuBuildSide = {
-    GpuJoinUtils.getGpuBuildSide(join.buildSide)
-  }
-}
