+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuSortMergeJoinMeta.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2021, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,39 +14,32 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims.v2
+package com.nvidia.spark.rapids
 
-import com.nvidia.spark.rapids._
-
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
-import org.apache.spark.sql.catalyst.plans.JoinType
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
+import org.apache.spark.sql.execution.SortExec
+import org.apache.spark.sql.execution.joins.SortMergeJoinExec
 import org.apache.spark.sql.rapids.execution.{GpuHashJoin, JoinTypeChecks}
 
-object GpuJoinUtils {
-  def getGpuBuildSide(buildSide: BuildSide): GpuBuildSide = {
-    buildSide match {
-      case BuildRight => GpuBuildRight
-      case BuildLeft => GpuBuildLeft
-      case _ => throw new Exception("unknown buildSide Type")
-    }
-  }
-}
-
-class GpuShuffledHashJoinMeta(
-    join: ShuffledHashJoinExec,
+class GpuSortMergeJoinMeta(
+    join: SortMergeJoinExec,
     conf: RapidsConf,
     parent: Option[RapidsMeta[_, _, _]],
     rule: DataFromReplacementRule)
-  extends SparkPlanMeta[ShuffledHashJoinExec](join, conf, parent, rule) {
+  extends SparkPlanMeta[SortMergeJoinExec](join, conf, parent, rule) {
+
   val leftKeys: Seq[BaseExprMeta[_]] =
     join.leftKeys.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
   val rightKeys: Seq[BaseExprMeta[_]] =
     join.rightKeys.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
-  val condition: Option[BaseExprMeta[_]] =
-    join.condition.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
+  val condition: Option[BaseExprMeta[_]] = join.condition.map(
+    GpuOverrides.wrapExpr(_, conf, Some(this)))
+  val buildSide: GpuBuildSide = if (GpuHashJoin.canBuildRight(join.joinType)) {
+    GpuBuildRight
+  } else if (GpuHashJoin.canBuildLeft(join.joinType)) {
+    GpuBuildLeft
+  } else {
+    throw new IllegalStateException(s"Cannot build either side for ${join.joinType} join")
+  }
 
   override val childExprs: Seq[BaseExprMeta[_]] = leftKeys ++ rightKeys ++ condition
 
@@ -54,7 +47,29 @@ class GpuShuffledHashJoinMeta(
     JoinTypeChecks.equiJoinMeta(leftKeys, rightKeys, condition)
 
   override def tagPlanForGpu(): Unit = {
+    // Use conditions from Hash Join
     GpuHashJoin.tagJoin(this, join.joinType, join.leftKeys, join.rightKeys, join.condition)
+
+    if (!conf.enableReplaceSortMergeJoin) {
+      willNotWorkOnGpu(s"Not replacing sort merge join with hash join, " +
+        s"see ${RapidsConf.ENABLE_REPLACE_SORTMERGEJOIN.key}")
+    }
+
+    // make sure this is the last check - if this is SortMergeJoin, the children can be Sorts and we
+    // want to validate they can run on GPU and remove them before replacing this with a
+    // ShuffleHashJoin
+    if (canThisBeReplaced) {
+      childPlans.foreach { plan =>
+        if (plan.wrapped.isInstanceOf[SortExec]) {
+          if (!plan.canThisBeReplaced) {
+            willNotWorkOnGpu(s"can't replace sortMergeJoin because one of the SortExec's before " +
+              s"can't be replaced.")
+          } else {
+            plan.shouldBeRemoved("replacing sortMergeJoin with shuffleHashJoin")
+          }
+        }
+      }
+    }
   }
 
   override def convertToGpu(): GpuExec = {
@@ -63,36 +78,15 @@ class GpuShuffledHashJoinMeta(
       leftKeys.map(_.convertToGpu()),
       rightKeys.map(_.convertToGpu()),
       join.joinType,
-      GpuJoinUtils.getGpuBuildSide(join.buildSide),
+      buildSide,
       None,
       left,
       right,
-      isSkewJoin = false)(
+      join.isSkewJoin)(
       join.leftKeys,
       join.rightKeys)
     // The GPU does not yet support conditional joins, so conditions are implemented
     // as a filter after the join when possible.
     condition.map(c => GpuFilterExec(c.convertToGpu(), joinExec)).getOrElse(joinExec)
   }
-}
-
-case class GpuShuffledHashJoinExec(
-    override val leftKeys: Seq[Expression],
-    override val rightKeys: Seq[Expression],
-    joinType: JoinType,
-    buildSide: GpuBuildSide,
-    override val condition: Option[Expression],
-    left: SparkPlan,
-    right: SparkPlan,
-    override val isSkewJoin: Boolean)(
-    cpuLeftKeys: Seq[Expression],
-    cpuRightKeys: Seq[Expression])
-  extends GpuShuffledHashJoinBase(
-    buildSide,
-    condition,
-    isSkewJoin = isSkewJoin,
-    cpuLeftKeys,
-    cpuRightKeys) {
-
-  override def otherCopyArgs: Seq[AnyRef] = cpuLeftKeys :: cpuRightKeys :: Nil
 }
