+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/GpuBroadcastNestedLoopJoinExec.scala
@@ -18,7 +18,7 @@ package org.apache.spark.sql.rapids.execution
 
 import ai.rapids.cudf.{ast, GatherMap, NvtxColor, OutOfBoundsPolicy, Table}
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.shims.v2.ShimBinaryExecNode
+import com.nvidia.spark.rapids.shims.v2.{GpuJoinUtils, ShimBinaryExecNode}
 
 import org.apache.spark.TaskContext
 import org.apache.spark.broadcast.Broadcast
@@ -43,13 +43,14 @@ class GpuBroadcastNestedLoopJoinMeta(
   val conditionMeta: Option[BaseExprMeta[_]] =
     join.condition.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
 
+  val gpuBuildSide: GpuBuildSide = GpuJoinUtils.getGpuBuildSide(join.buildSide)
+
   override def namedChildExprs: Map[String, Seq[BaseExprMeta[_]]] =
     JoinTypeChecks.nonEquiJoinMeta(conditionMeta)
 
   override val childExprs: Seq[BaseExprMeta[_]] = conditionMeta.toSeq
 
   override def tagPlanForGpu(): Unit = {
-    val gpuBuildSide = ShimLoader.getSparkShims.getBuildSide(join)
     JoinTypeChecks.tagForGpu(join.joinType, this)
     join.joinType match {
       case _: InnerLike =>
@@ -84,7 +85,6 @@ class GpuBroadcastNestedLoopJoinMeta(
   override def convertToGpu(): GpuExec = {
     val Seq(left, right) = childPlans.map(_.convertIfNeeded())
     // The broadcast part of this must be a BroadcastExchangeExec
-    val gpuBuildSide = ShimLoader.getSparkShims.getBuildSide(join)
     val buildSide = gpuBuildSide match {
       case GpuBuildLeft => left
       case GpuBuildRight => right
@@ -105,9 +105,9 @@ class GpuBroadcastNestedLoopJoinMeta(
       case _ => throw new IllegalStateException(s"Unsupported join type ${join.joinType}")
     }
 
-    val joinExec = ShimLoader.getSparkShims.getGpuBroadcastNestedLoopJoinShim(
-      left, right, join,
-      join.joinType,
+    val joinExec = GpuBroadcastNestedLoopJoinExec(
+      left, right,
+      join.joinType, gpuBuildSide,
       if (isAstCondition) condition else None,
       conf.gpuTargetBatchSizeBytes)
     if (isAstCondition) {
@@ -300,7 +300,7 @@ class ConditionalNestedLoopJoinIterator(
   }
 }
 
-object GpuBroadcastNestedLoopJoinExecBase extends Arm {
+object GpuBroadcastNestedLoopJoinExec extends Arm {
   def nestedLoopJoin(
       joinType: JoinType,
       buildSide: GpuBuildSide,
@@ -367,18 +367,15 @@ object GpuBroadcastNestedLoopJoinExecBase extends Arm {
   }
 }
 
-abstract class GpuBroadcastNestedLoopJoinExecBase(
+case class GpuBroadcastNestedLoopJoinExec(
     left: SparkPlan,
     right: SparkPlan,
     joinType: JoinType,
+    gpuBuildSide: GpuBuildSide,
     condition: Option[Expression],
     targetSizeBytes: Long) extends ShimBinaryExecNode with GpuExec {
 
   import GpuMetric._
-
-  // Spark BuildSide, BuildRight, BuildLeft changed packages between Spark versions
-  // so return a GPU version that is agnostic to the Spark version.
-  def getGpuBuildSide: GpuBuildSide
 
   override protected def doExecute(): RDD[InternalRow] =
     throw new IllegalStateException("This should only be called from columnar")
@@ -393,7 +390,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
     JOIN_OUTPUT_ROWS -> createMetric(MODERATE_LEVEL, DESCRIPTION_JOIN_OUTPUT_ROWS)) ++ spillMetrics
 
   /** BuildRight means the right relation <=> the broadcast relation. */
-  private val (streamed, broadcast) = getGpuBuildSide match {
+  private val (streamed, broadcast) = gpuBuildSide match {
     case GpuBuildRight => (left, right)
     case GpuBuildLeft => (right, left)
   }
@@ -407,7 +404,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
     case reused: ReusedExchangeExec => reused.child.asInstanceOf[GpuBroadcastExchangeExecBase]
   }
 
-  override def requiredChildDistribution: Seq[Distribution] = getGpuBuildSide match {
+  override def requiredChildDistribution: Seq[Distribution] = gpuBuildSide match {
     case GpuBuildLeft =>
       BroadcastDistribution(IdentityBroadcastMode) :: UnspecifiedDistribution :: Nil
     case GpuBuildRight =>
@@ -474,8 +471,8 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
       // For outer joins use a true condition if there are any columns in the build side
       // otherwise use a cross join.
       val useTrueCondition = joinType match {
-        case LeftOuter if getGpuBuildSide == GpuBuildRight => right.output.nonEmpty
-        case RightOuter if getGpuBuildSide == GpuBuildLeft => left.output.nonEmpty
+        case LeftOuter if gpuBuildSide == GpuBuildRight => right.output.nonEmpty
+        case RightOuter if gpuBuildSide == GpuBuildLeft => left.output.nonEmpty
         case _ => false
       }
       if (useTrueCondition) Some(GpuLiteral(true)) else None
@@ -493,7 +490,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
       exists: Boolean,
       buildTime: GpuMetric,
       buildDataSize: GpuMetric): RDD[ColumnarBatch] = {
-    assert(getGpuBuildSide == GpuBuildRight)
+    assert(gpuBuildSide == GpuBuildRight)
     streamed.executeColumnar().mapPartitionsInternal { streamedIter =>
       val buildRows = computeBuildRowCount(broadcastRelation, buildTime, buildDataSize)
       if (buildRows > 0 == exists) {
@@ -517,13 +514,13 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
       lazy val builtBatch = makeBuiltBatch(broadcastRelation, buildTime, buildDataSize)
       val joinIterator: RDD[ColumnarBatch] = joinType match {
         case LeftSemi =>
-          if (getGpuBuildSide == GpuBuildRight) {
+          if (gpuBuildSide == GpuBuildRight) {
             leftExistenceJoin(broadcastRelation, exists=true, buildTime, buildDataSize)
           } else {
             left.executeColumnar()
           }
         case LeftAnti =>
-          if (getGpuBuildSide == GpuBuildRight) {
+          if (gpuBuildSide == GpuBuildRight) {
             leftExistenceJoin(broadcastRelation, exists=false, buildTime, buildDataSize)
           } else {
             // degenerate case, no rows are returned.
@@ -532,7 +529,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
           }
         case _ =>
           // Everything else is treated like an unconditional cross join
-          val buildSide = getGpuBuildSide
+          val buildSide = gpuBuildSide
           val spillCallback = GpuMetric.makeSpillCallback(allMetrics)
           val joinTime = gpuLongMetric(JOIN_TIME)
           streamed.executeColumnar().mapPartitions { streamedIter =>
@@ -590,7 +587,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
       val numOutputBatches = gpuLongMetric(NUM_OUTPUT_BATCHES)
       val semWait = gpuLongMetric(SEMAPHORE_WAIT_TIME)
       val counts = streamed.executeColumnar().map(getRowCountAndClose)
-      GpuBroadcastNestedLoopJoinExecBase.divideIntoBatches(
+      GpuBroadcastNestedLoopJoinExec.divideIntoBatches(
         counts.map(s => s * buildCount),
         targetSizeBytes,
         numOutputRows,
@@ -614,7 +611,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
     val joinTime = gpuLongMetric(JOIN_TIME)
     val joinOutputRows = gpuLongMetric(JOIN_OUTPUT_ROWS)
     val nestedLoopJoinType = joinType
-    val buildSide = getGpuBuildSide
+    val buildSide = gpuBuildSide
     streamed.executeColumnar().mapPartitions { streamedIter =>
       val lazyStream = streamedIter.map { cb =>
         withResource(cb) { cb =>
@@ -625,7 +622,7 @@ abstract class GpuBroadcastNestedLoopJoinExecBase(
         LazySpillableColumnarBatch(_, spillCallback, "built_batch")
       }
 
-      GpuBroadcastNestedLoopJoinExecBase.nestedLoopJoin(
+      GpuBroadcastNestedLoopJoinExec.nestedLoopJoin(
         nestedLoopJoinType, buildSide, numFirstTableColumns,
         spillableBuiltBatch,
         lazyStream, streamAttributes, targetSizeBytes, boundCondition, spillCallback,
